Project envs cleaning.

+++ corr-db/corrdb/common/managers/storage_manager.py
@@ -192,13 +192,12 @@ class StorageManager:
         for environment_id in project.history:
             _environment = EnvironmentModel.objects.with_id(environment_id)
             if _environment:
-                if _environment.bundle.scope == "local":
+                if _environment.bundle and _environment.bundle.scope == "local":
                     result = self.storage_delete_file('bundle', _environment.bundle.location)
                     if result:
                         logStat(deleted=True, bundle=_environment.bundle)
                         logStat(deleted=True, environment=_environment)
                         _environment.bundle.delete()
-                        _environment.delete()
                 else:
                     logStat(deleted=True, environment=_environment)
                 _environment.delete()
